Enabling bigger read/write buffering, without loading the full file into memory. Hopefully this will be a happy medium for speed and virt size.
Moving many imports to the start of the module for functions and methods
that are called many times in a normal run.

Adding docstrings and fixing variable names to come up to convention
standards.

Fixing issue where we were unpacking data that was just thrown away
immediately afterwards; we won't unpack it anymore.