Fixed bug when reading manifest from Jar
An UnsupportedOperation exception is thrown when get_manifest is called on a
JarInfo object.

This exception occurs because get_manifest passes a zipfile.ZipExtFile object
to the parse method of the Manifest class. The ZipExtFile class does not
support tell and seek operations, and therefore throws the UnsupportedOperation
exception.

To fix this, get_manifest reads the zipfile.ZipExtFile object and passes a
string to Manifest's parse method.