Fixed error in verification logic.
Commit f88f390 (refactoring javatools/jarutil.py:verify()) introduced an error
in the logic of signature manifest verification:

    if not signature_manifest.verify_manifest_main_checksum(jar_manifest):
        # The above is allowed to fail. If so, second attempt below:
        errors = signature_manifest.verify_manifest_entry_checksums(jar_manifest)
        if len(errors) > 0:
            return ("error message")

The newly introduced verify_manifest_main_checksum() is not the function which
is allowed to fail. It has misleading name: it checks first the main checksum,
and then the checksum of the "main attributes". If the first fails, the second
is not allowed to fail.

Fixed by splitting the function into two and correctly using the parts.

Closes #26.