Python 3 compatible syntax in __init__.py
In Python 3, buffer (i.e. memoryview) does not have a constructor
with offset, on which code of _next_argsig() relies.
Change _next_argsig() to work on strings instead.
This does not impact performance as this function is used on Java
identifiers, which are small in size.

No external interfaces are changed.