Loading manifest from JAR factored into own function
When JAR manifest is read from JAR file, that file is zip. The returned data
is 'str' in Python 2 and 'bytes' in Python 3; by its nature, the manifest
content is UTF-8 encoded text. This requires different processing in Py2/Py3.

This processing is factored in own function. Loading manifest from JAR file
is done only with this new function. It also makes redundant
Jarinfo:get_manifest() (which was not covered by tests anyway).